split the chonky tests up
