+++ e2e/test/scenarios/dashboard-filters-2/dashboard-filters-2-stage-2.cy.spec.ts
@@ -1,0 +1,374 @@
+import { H } from "e2e/support";
+
+import * as QSHelpers from "./shared/dashboard-filters-query-stages";
+
+/**
+ * Empty section title element is rendered.
+ * TODO: https://github.com/metabase/metabase/issues/47218
+ */
+const NAMELESS_SECTION = "";
+
+/**
+ * Abbreviations used for card aliases in this test suite:
+ *  qbq = question-based question
+ *  qbm = question-based model
+ *  mbq = model-based question
+ *  mbm = model-based model
+ */
+describe("scenarios > dashboard > filters > query stages", () => {
+  beforeEach(() => {
+    H.restore();
+    cy.signInAsAdmin();
+    QSHelpers.createBaseQuestions();
+
+    cy.intercept("POST", "/api/dataset").as("dataset");
+    cy.intercept("GET", "/api/dashboard/**").as("getDashboard");
+    cy.intercept("PUT", "/api/dashboard/**").as("updateDashboard");
+    cy.intercept("POST", "/api/dashboard/*/dashcard/*/card/*/query").as(
+      "dashboardData",
+    );
+    cy.intercept("GET", "/api/public/dashboard/*/dashcard/*/card/*").as(
+      "publicDashboardData",
+    );
+    cy.intercept("GET", "/api/embed/dashboard/*/dashcard/*/card/*").as(
+      "embeddedDashboardData",
+    );
+  });
+
+  describe("2-stage queries", () => {
+    describe("Q6 - Q4 + 2nd stage with join, custom column, 2 aggregations, no breakouts", () => {
+      beforeEach(() => {
+        QSHelpers.createAndVisitDashboardWithCardMatrix(
+          QSHelpers.createQ6Query,
+        );
+      });
+
+      it("allows to map to all relevant columns", () => {
+        H.editDashboard();
+
+        cy.log("## date columns");
+        QSHelpers.getFilter("Date").click();
+        verifyDateMappingOptions();
+
+        cy.log("## text columns");
+        QSHelpers.getFilter("Text").click();
+        verifyTextMappingOptions();
+
+        cy.log("## number columns");
+        QSHelpers.getFilter("Number").click();
+        verifyNumberMappingOptions();
+
+        function verifyDateMappingOptions() {
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.QUESTION_BASED_QUESTION_INDEX,
+            [
+              ["Base Orders Question", QSHelpers.ORDERS_DATE_COLUMNS],
+              [
+                "Reviews",
+                [
+                  ...QSHelpers.REVIEWS_DATE_COLUMNS,
+                  ...QSHelpers.REVIEWS_DATE_COLUMNS,
+                ],
+              ], // TODO: https://github.com/metabase/metabase/issues/46845
+              [
+                "Product",
+                [
+                  ...QSHelpers.PRODUCTS_DATE_COLUMNS,
+                  ...QSHelpers.PRODUCTS_DATE_COLUMNS,
+                ],
+              ], // TODO: https://github.com/metabase/metabase/issues/46845
+              ["User", QSHelpers.PEOPLE_DATE_COLUMNS],
+              ["Summaries", ["Created At: Month", "Created At: Year"]],
+            ],
+          );
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.MODEL_BASED_QUESTION_INDEX,
+            [
+              ["Base Orders Model", QSHelpers.ORDERS_DATE_COLUMNS],
+              [
+                "Reviews",
+                [
+                  ...QSHelpers.REVIEWS_DATE_COLUMNS,
+                  ...QSHelpers.REVIEWS_DATE_COLUMNS,
+                ],
+              ], // TODO: https://github.com/metabase/metabase/issues/46845
+              [
+                "Product",
+                [
+                  ...QSHelpers.PRODUCTS_DATE_COLUMNS,
+                  ...QSHelpers.PRODUCTS_DATE_COLUMNS,
+                ],
+              ], // TODO: https://github.com/metabase/metabase/issues/46845
+              ["User", QSHelpers.PEOPLE_DATE_COLUMNS],
+              ["Summaries", ["Created At: Month", "Created At: Year"]],
+            ],
+          );
+          QSHelpers.verifyNoDashcardMappingOptions(
+            QSHelpers.QUESTION_BASED_MODEL_INDEX,
+          );
+          QSHelpers.verifyNoDashcardMappingOptions(
+            QSHelpers.MODEL_BASED_MODEL_INDEX,
+          );
+        }
+
+        function verifyTextMappingOptions() {
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.QUESTION_BASED_QUESTION_INDEX,
+            [
+              [
+                "Reviews",
+                [
+                  ...QSHelpers.REVIEWS_TEXT_COLUMNS,
+                  ...QSHelpers.REVIEWS_TEXT_COLUMNS,
+                ],
+              ],
+              [
+                "Product",
+                [
+                  ...QSHelpers.PRODUCTS_TEXT_COLUMNS,
+                  ...QSHelpers.PRODUCTS_TEXT_COLUMNS,
+                ],
+              ], // TODO: https://github.com/metabase/metabase/issues/46845
+              ["User", QSHelpers.PEOPLE_TEXT_COLUMNS],
+              ["Summaries", ["Category"]],
+            ],
+          );
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.MODEL_BASED_QUESTION_INDEX,
+            [
+              [
+                "Reviews",
+                [
+                  ...QSHelpers.REVIEWS_TEXT_COLUMNS,
+                  ...QSHelpers.REVIEWS_TEXT_COLUMNS,
+                ],
+              ],
+              [
+                "Product",
+                [
+                  ...QSHelpers.PRODUCTS_TEXT_COLUMNS,
+                  ...QSHelpers.PRODUCTS_TEXT_COLUMNS,
+                ],
+              ], // TODO: https://github.com/metabase/metabase/issues/46845
+              ["User", QSHelpers.PEOPLE_TEXT_COLUMNS],
+              ["Summaries", ["Category"]],
+            ],
+          );
+          QSHelpers.verifyNoDashcardMappingOptions(
+            QSHelpers.QUESTION_BASED_MODEL_INDEX,
+          );
+          QSHelpers.verifyNoDashcardMappingOptions(
+            QSHelpers.MODEL_BASED_MODEL_INDEX,
+          );
+        }
+
+        function verifyNumberMappingOptions() {
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.QUESTION_BASED_QUESTION_INDEX,
+            [
+              [
+                "Base Orders Question",
+                [...QSHelpers.ORDERS_NUMBER_COLUMNS, "Net"],
+              ],
+              [
+                "Reviews",
+                [
+                  ...QSHelpers.REVIEWS_NUMBER_COLUMNS,
+                  ...QSHelpers.REVIEWS_NUMBER_COLUMNS,
+                ],
+              ],
+              [
+                "Product",
+                [
+                  ...QSHelpers.PRODUCTS_NUMBER_COLUMNS,
+                  ...QSHelpers.PRODUCTS_NUMBER_COLUMNS,
+                ], // TODO: https://github.com/metabase/metabase/issues/46845
+              ],
+              ["User", QSHelpers.PEOPLE_NUMBER_COLUMNS],
+              ["Summaries", ["Count", "Sum of Total", "5 * Count"]],
+            ],
+          );
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.MODEL_BASED_QUESTION_INDEX,
+            [
+              [
+                "Base Orders Model",
+                [...QSHelpers.ORDERS_NUMBER_COLUMNS, "Net"],
+              ],
+              [
+                "Reviews",
+                [
+                  ...QSHelpers.REVIEWS_NUMBER_COLUMNS,
+                  ...QSHelpers.REVIEWS_NUMBER_COLUMNS,
+                ],
+              ],
+              [
+                "Product",
+                [
+                  ...QSHelpers.PRODUCTS_NUMBER_COLUMNS,
+                  ...QSHelpers.PRODUCTS_NUMBER_COLUMNS,
+                ], // TODO: https://github.com/metabase/metabase/issues/46845
+              ],
+              ["User", QSHelpers.PEOPLE_NUMBER_COLUMNS],
+              ["Summaries", ["Count", "Sum of Total", "5 * Count"]],
+            ],
+          );
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.QUESTION_BASED_MODEL_INDEX,
+            [
+              [
+                NAMELESS_SECTION,
+                ["Count", "Sum of Reviews - Created At: Month → Rating"],
+              ],
+            ],
+          );
+          QSHelpers.verifyDashcardMappingOptions(
+            QSHelpers.MODEL_BASED_MODEL_INDEX,
+            [
+              [
+                NAMELESS_SECTION,
+                ["Count", "Sum of Reviews - Created At: Month → Rating"],
+              ],
+            ],
+          );
+        }
+      });
+
+      describe("applies filter to the the dashcard and allows to drill via dashcard header", () => {
+        it("1st stage explicit join", () => {
+          QSHelpers.setup1stStageExplicitJoinFilter();
+          QSHelpers.apply1stStageExplicitJoinFilter();
+          cy.wait(["@dashboardData", "@dashboardData"]);
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["1,813", "7,218"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["1,813", "7,218"],
+          });
+        });
+
+        it("1st stage implicit join (data source)", () => {
+          QSHelpers.setup1stStageImplicitJoinFromSourceFilter();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["2,071", "8,252"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["2,071", "8,252"],
+          });
+        });
+
+        // TODO: https://github.com/metabase/metabase/issues/46774
+        it.skip("1st stage implicit join (joined data source)", () => {
+          QSHelpers.setup1stStageImplicitJoinFromJoinFilter();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["4,447", "17,714"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["4,447", "17,714"],
+          });
+        });
+
+        it("1st stage custom column", () => {
+          QSHelpers.setup1stStageCustomColumnFilter();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["971", "3,900"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["971", "3,900"],
+          });
+        });
+
+        it("1st stage aggregation", () => {
+          QSHelpers.setup1stStageAggregationFilter();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["3", "13"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["3", "13"],
+          });
+        });
+
+        // TODO: https://github.com/metabase/metabase/issues/46774
+        it.skip("1st stage breakout", () => {
+          QSHelpers.setup1stStageBreakoutFilter();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["4,447", "17,714"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["4,447", "17,714"],
+          });
+        });
+
+        it("2nd stage explicit join", () => {
+          QSHelpers.setup2ndStageExplicitJoinFilter();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["16", "80"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["16", "80"],
+          });
+        });
+
+        it("2nd stage custom column", () => {
+          QSHelpers.setup2ndStageCustomColumnFilter();
+          QSHelpers.apply2ndStageCustomColumnFilter();
+          cy.wait(["@dashboardData", "@dashboardData"]);
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 0,
+            values: ["31", "114"],
+          });
+
+          QSHelpers.goBackToDashboard();
+
+          QSHelpers.verifyDashcardCellValues({
+            dashcardIndex: 1,
+            values: ["31", "114"],
+          });
+        });
+      });
+    });
+  });
+});
